[feature] Weed out unreferenced attachments
As it turns out, copying a page in Confluence® copies all the
attachments with it behind the scenes; and these survive into the ZIP
export somehow, regardless of whether people (thought they) edited the
copy to remove the attachment!

- New `exclude_unused_attachments()` function on the R side
- The Python side can now `.list_attachments()` out of the `ri:filename`
XML attributes. As a courtesy to the R side, the return value of
`.list_attachments()` is keyed with the `latest.version` of the
documents (which obviously Python now needs to keep track of as well).
This spares R an extra left join, which in turn would require
some unsightly passing around of the documents tibble, despite the
DocumentConverter being “supposed to know” about that data already.
(Cross-language, cross-programming-paradigm DRY is trickier than
I thought...)
- The whole change sheds a lot of dead weight obviously, and brings
the size of the ISAS-FSD Outline import way below the 4G
limit. Switch back to ZIP_32 format, as Outline won't parse ZIP_64
at all anyway (symptom is [`invalid zip64 end of central directory
locator signature`](https://github.com/thejoshwolfe/yauzl?tab=readme-ov-file#change-history)
in the logs)